rename package shared to pkg
