+++ images/coverage-go112/Makefile
@@ -13,4 +13,4 @@
 # limitations under the License.
 
 IMAGE_NAME ?= coverage-go112
-include ../../shared/Makefile.simple-image
+include ../../pkg/Makefile.simple-image
