+++ images/flaky-test-reporter/Makefile
@@ -13,4 +13,4 @@
 # limitations under the License.
 
 IMAGE_NAME = flaky-test-reporter
-include ../../shared/Makefile.simple-image
+include ../../pkg/Makefile.simple-image
