+++ images/prow-tests/Makefile
@@ -13,4 +13,4 @@
 # limitations under the License.
 
 IMAGE_NAME = prow-tests
-include ../../shared/Makefile.simple-image
+include ../../pkg/Makefile.simple-image
