+++ pkg/Makefile.simple-image
