+++ pkg/README.md
@@ -1,0 +1,3 @@
+# Test Infrastructure packages
+
+This directory contains shared packages used by tools in this repository.
