+++ pkg/git/info.go
