+++ pkg/testgrid/testgrid.go
@@ -24,9 +24,9 @@ import (
 	"os"
 	"path"
 
-	"knative.dev/test-infra/shared/common"
-	"knative.dev/test-infra/shared/junit"
-	"knative.dev/test-infra/shared/prow"
+	"knative.dev/test-infra/pkg/common"
+	"knative.dev/test-infra/pkg/junit"
+	"knative.dev/test-infra/pkg/prow"
 )
 
 const (
