--- shared/README.md
@@ -1,3 +1,0 @@
-# Test Infrastructure shared packages
-
-This directory contains shared packages used by other repositories.
