+++ tools/flaky-test-reporter/result.go
@@ -28,9 +28,9 @@ import (
 	"sort"
 	"strings"
 
-	"knative.dev/test-infra/shared/common"
-	"knative.dev/test-infra/shared/junit"
-	"knative.dev/test-infra/shared/prow"
+	"knative.dev/test-infra/pkg/common"
+	"knative.dev/test-infra/pkg/junit"
+	"knative.dev/test-infra/pkg/prow"
 	"knative.dev/test-infra/tools/flaky-test-reporter/config"
 )
