+++ tools/flaky-test-retryer/log_parser.go
@@ -25,8 +25,8 @@ import (
 	"path/filepath"
 	"strings"
 
-	"knative.dev/test-infra/shared/junit"
-	"knative.dev/test-infra/shared/prow"
+	"knative.dev/test-infra/pkg/junit"
+	"knative.dev/test-infra/pkg/prow"
 	"knative.dev/test-infra/tools/flaky-test-reporter/jsonreport"
 
 	// TODO: remove this import once "k8s.io/test-infra" import problems are fixed
